Refactor build command ownership
