+++ erun-common/build_spec.go
@@ -1,0 +1,233 @@
+package eruncommon
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+)
+
+func ResolveDockerImageReference(store DockerStore, findProjectRoot ProjectFinderFunc, resolveBuildContext BuildContextResolverFunc, now NowFunc, buildDir string, target DockerCommandTarget) (DockerImageReference, error) {
+	store, findProjectRoot, _, now = normalizeDockerDependencies(store, findProjectRoot, resolveBuildContext, now)
+
+	target, _, err := ResolveDockerBuildTarget(findProjectRoot, target)
+	if err != nil {
+		return DockerImageReference{}, err
+	}
+
+	projectRoot, err := resolveDockerBuildProjectRoot(findProjectRoot, target)
+	if err != nil {
+		return DockerImageReference{}, err
+	}
+
+	environment, err := resolveDockerBuildEnvironment(store, findProjectRoot, projectRoot, target.Environment)
+	if err != nil {
+		return DockerImageReference{}, err
+	}
+
+	return resolveDockerImageReferenceForProject(now, projectRoot, environment, buildDir, strings.TrimSpace(target.VersionOverride))
+}
+
+func ResolveDockerBuildForComponent(store DockerStore, findProjectRoot ProjectFinderFunc, resolveBuildContext BuildContextResolverFunc, now NowFunc, projectRoot, environment, componentName, versionOverride string) (*DockerBuildSpec, error) {
+	_, _, resolveBuildContext, now = normalizeDockerDependencies(store, findProjectRoot, resolveBuildContext, now)
+
+	if !isLocalEnvironment(environment) {
+		return nil, nil
+	}
+
+	if buildContext, ok := currentComponentDockerBuildContext(resolveBuildContext, componentName); ok {
+		build, err := newDockerBuildSpec(now, projectRoot, environment, buildContext, versionOverride)
+		if err != nil {
+			return nil, err
+		}
+		return &build, nil
+	}
+
+	buildContext, ok, err := FindComponentDockerBuildContext(projectRoot, componentName)
+	if err != nil || !ok {
+		return nil, err
+	}
+
+	build, err := newDockerBuildSpec(now, projectRoot, environment, buildContext, versionOverride)
+	if err != nil {
+		return nil, err
+	}
+	return &build, nil
+}
+
+func currentComponentDockerBuildContext(resolveBuildContext BuildContextResolverFunc, componentName string) (DockerBuildContext, bool) {
+	if resolveBuildContext == nil {
+		return DockerBuildContext{}, false
+	}
+	buildContext, err := resolveBuildContext()
+	if err != nil {
+		return DockerBuildContext{}, false
+	}
+	dir := filepath.Clean(strings.TrimSpace(buildContext.Dir))
+	if filepath.Base(dir) != strings.TrimSpace(componentName) || filepath.Base(filepath.Dir(dir)) != "docker" {
+		return DockerBuildContext{}, false
+	}
+	return buildContext, strings.TrimSpace(buildContext.DockerfilePath) != ""
+}
+
+func ResolveDockerBuildForImageReference(store DockerStore, findProjectRoot ProjectFinderFunc, resolveBuildContext BuildContextResolverFunc, now NowFunc, projectRoot, environment, image string) (DockerBuildSpec, bool, error) {
+	image = strings.TrimSpace(image)
+	if image == "" {
+		return DockerBuildSpec{}, false, nil
+	}
+
+	nameTag := image
+	registry := ""
+	if idx := strings.LastIndex(image, "/"); idx >= 0 {
+		registry = image[:idx]
+		nameTag = image[idx+1:]
+	}
+
+	imageName, version, ok := strings.Cut(nameTag, ":")
+	if !ok || strings.TrimSpace(imageName) == "" || strings.TrimSpace(version) == "" {
+		return DockerBuildSpec{}, false, nil
+	}
+
+	buildContext, ok, err := FindComponentDockerBuildContext(projectRoot, imageName)
+	if err != nil || !ok {
+		return DockerBuildSpec{}, false, err
+	}
+
+	imageRef := DockerImageReference{
+		ProjectRoot:  projectRoot,
+		Environment:  strings.TrimSpace(environment),
+		Registry:     registry,
+		ImageName:    imageName,
+		Version:      version,
+		Tag:          image,
+		IsLocalBuild: isLocalEnvironment(environment),
+	}
+	skipIfExists, err := resolveDockerBuildSkipIfExists(projectRoot, environment, imageRef)
+	if err != nil {
+		return DockerBuildSpec{}, false, err
+	}
+
+	return DockerBuildSpec{
+		ContextDir:     ResolveDockerBuildContextDirForProject(buildContext.Dir, projectRoot),
+		DockerfilePath: buildContext.DockerfilePath,
+		Image:          imageRef,
+		SkipIfExists:   skipIfExists,
+	}, true, nil
+}
+
+func resolveDockerBuildSpec(store DockerStore, findProjectRoot ProjectFinderFunc, resolveBuildContext BuildContextResolverFunc, now NowFunc, buildContext DockerBuildContext, target DockerCommandTarget) (DockerBuildSpec, error) {
+	projectRoot, err := resolveDockerBuildProjectRoot(findProjectRoot, target)
+	if err != nil {
+		return DockerBuildSpec{}, err
+	}
+
+	environment, err := resolveDockerBuildEnvironment(store, findProjectRoot, projectRoot, target.Environment)
+	if err != nil {
+		return DockerBuildSpec{}, err
+	}
+
+	return newDockerBuildSpec(now, projectRoot, environment, buildContext, strings.TrimSpace(target.VersionOverride))
+}
+
+func resolveDockerImageReferenceForProject(now NowFunc, projectRoot, environment, buildDir, versionOverride string) (DockerImageReference, error) {
+	registry, err := resolveDockerBuildRegistryForEnvironment(projectRoot, environment)
+	if err != nil {
+		return DockerImageReference{}, err
+	}
+
+	imageName := strings.TrimSpace(filepath.Base(buildDir))
+	if imageName == "" || imageName == "." || imageName == string(filepath.Separator) {
+		return DockerImageReference{}, fmt.Errorf("could not determine image name from current directory")
+	}
+
+	version, versionFromBuildDir, versionFilePath, err := resolveDockerImageVersion(now, projectRoot, environment, buildDir, versionOverride)
+	if err != nil {
+		return DockerImageReference{}, err
+	}
+
+	return DockerImageReference{
+		ProjectRoot:         projectRoot,
+		Environment:         strings.TrimSpace(environment),
+		Registry:            registry,
+		ImageName:           imageName,
+		Version:             version,
+		Tag:                 fmt.Sprintf("%s/%s:%s", strings.TrimRight(registry, "/"), imageName, version),
+		IsLocalBuild:        isLocalEnvironment(environment),
+		VersionFilePath:     versionFilePath,
+		VersionFromBuildDir: versionFromBuildDir,
+	}, nil
+}
+
+func resolveDockerImageVersion(now NowFunc, projectRoot, environment, buildDir, versionOverride string) (string, bool, string, error) {
+	baseVersion, versionFromBuildDir, versionFilePath, err := ResolveDockerBuildVersion(buildDir, projectRoot)
+	if err != nil {
+		return "", false, "", err
+	}
+
+	if versionOverride = strings.TrimSpace(versionOverride); versionOverride != "" {
+		if versionFromBuildDir {
+			return baseVersion, versionFromBuildDir, versionFilePath, nil
+		}
+		return versionOverride, versionFromBuildDir, versionFilePath, nil
+	}
+
+	if !isLocalEnvironment(environment) || versionFromBuildDir {
+		return baseVersion, versionFromBuildDir, versionFilePath, nil
+	}
+	return formatLocalSnapshotVersion(baseVersion, now()), versionFromBuildDir, versionFilePath, nil
+}
+
+func newDockerBuildSpec(now NowFunc, projectRoot, environment string, buildContext DockerBuildContext, versionOverride string) (DockerBuildSpec, error) {
+	if strings.TrimSpace(buildContext.DockerfilePath) == "" {
+		var err error
+		buildContext, err = DockerBuildContextAtDir(buildContext.Dir)
+		if err != nil {
+			return DockerBuildSpec{}, err
+		}
+	}
+
+	contextDir := ResolveDockerBuildContextDirForProject(buildContext.Dir, projectRoot)
+	imageRef, err := resolveDockerImageReferenceForProject(now, projectRoot, environment, buildContext.Dir, versionOverride)
+	if err != nil {
+		return DockerBuildSpec{}, err
+	}
+	skipIfExists, err := resolveDockerBuildSkipIfExists(projectRoot, environment, imageRef)
+	if err != nil {
+		return DockerBuildSpec{}, err
+	}
+
+	return DockerBuildSpec{
+		ContextDir:     contextDir,
+		DockerfilePath: buildContext.DockerfilePath,
+		Image:          imageRef,
+		SkipIfExists:   skipIfExists,
+	}, nil
+}
+
+func (b DockerBuildSpec) command() commandSpec {
+	args := dockerBuildArgs(b)
+	return commandSpec{
+		Dir:  b.ContextDir,
+		Name: "docker",
+		Args: args,
+	}
+}
+
+func (b DockerBuildSpec) traceCommands() []commandSpec {
+	if len(b.Platforms) == 0 {
+		return []commandSpec{b.command()}
+	}
+
+	return append(dockerBuildxSetupCommands(b.ContextDir), b.command())
+}
+
+func (p DockerPushSpec) command() commandSpec {
+	return commandSpec{
+		Dir:  p.Dir,
+		Name: "docker",
+		Args: []string{"push", p.Image.Tag},
+	}
+}
+
+func NewDockerPushSpec(dir string, image DockerImageReference) DockerPushSpec {
+	return DockerPushSpec{Dir: dir, Image: image}
+}
