+++ erun-common/build_types.go
@@ -1,0 +1,113 @@
+package eruncommon
+
+import (
+	"errors"
+	"io"
+	"time"
+)
+
+const localSnapshotTimestampFormat = "20060102150405"
+
+const multiPlatformBuildxBuilderName = "erun-multiarch"
+
+var (
+	ErrVersionFileNotFound        = errors.New("version file not found for current module")
+	ErrDockerBuildContextNotFound = errors.New("dockerfile not found in current directory")
+	ErrLinuxPackageBuildNotFound  = errors.New("linux package build script not found in current directory")
+	multiPlatformDockerBuilds     = []string{"linux/amd64", "linux/arm64"}
+)
+
+type commandSpec struct {
+	Dir  string   `json:"dir,omitempty"`
+	Name string   `json:"name"`
+	Args []string `json:"args"`
+}
+
+type (
+	BuildContextResolverFunc func() (DockerBuildContext, error)
+	NowFunc                  func() time.Time
+	DockerImageBuilderFunc   func(DockerBuildSpec, io.Writer, io.Writer) error
+	DockerImagePusherFunc    func(string, io.Writer, io.Writer) error
+	DockerImageInspectorFunc func(string) (bool, error)
+	DockerRegistryLoginFunc  func(string, io.Reader, io.Writer, io.Writer) error
+	BuildScriptRunnerFunc    func(string, string, []string, io.Reader, io.Writer, io.Writer) error
+	DockerPushFunc           func(Context, DockerPushSpec) error
+)
+
+type DockerStore interface {
+	ListTenantConfigs() ([]TenantConfig, error)
+	LoadTenantConfig(string) (TenantConfig, string, error)
+}
+
+type DockerBuildContext struct {
+	Dir            string
+	DockerfilePath string
+}
+
+type DockerImageReference struct {
+	ProjectRoot         string
+	Environment         string
+	Registry            string
+	ImageName           string
+	Version             string
+	Tag                 string
+	IsLocalBuild        bool
+	VersionFilePath     string
+	VersionFromBuildDir bool
+}
+
+type DockerBuildSpec struct {
+	ContextDir     string
+	DockerfilePath string
+	Image          DockerImageReference
+	Platforms      []string
+	Push           bool
+	SkipIfExists   bool
+}
+
+type DockerPushSpec struct {
+	Dir   string
+	Image DockerImageReference
+}
+
+type scriptSpec struct {
+	Dir  string
+	Path string
+	Env  []string
+}
+
+type BuildExecutionSpec struct {
+	release      *ReleaseSpec
+	script       *scriptSpec
+	linuxBuilds  []scriptSpec
+	dockerBuilds []DockerBuildSpec
+	dockerPushes []DockerPushSpec
+	skippedLinux bool
+}
+
+type DockerPushExecutionSpec struct {
+	builds []DockerBuildSpec
+	pushes []DockerPushSpec
+}
+
+type DockerCommandTarget struct {
+	ProjectRoot     string
+	Environment     string
+	VersionOverride string
+	Release         bool
+	Force           bool
+	Deploy          bool
+}
+
+type DockerRegistryAuthError struct {
+	Tag      string
+	Registry string
+	Message  string
+	Err      error
+}
+
+type LinuxPackageContext struct {
+	Dir               string
+	BuildScriptPath   string
+	ReleaseScriptPath string
+}
